LPS-57049 Move helper methods to the helper util
