+++ portal-impl/src/com/liferay/portlet/exportimport/lar/BridgeExportImportHelperImpl.java
@@ -86,6 +86,22 @@ public class BridgeExportImportHelperImpl implements ExportImportHelper {
 			portletRequest, paramPrefix, timeZoneSensitive);
 	}
 
+	@Override
+	public List<Portlet> getDataSiteLevelPortlets(long companyId)
+		throws Exception {
+
+		return getExportImportHelper().getDataSiteLevelPortlets(companyId);
+	}
+
+	@Override
+	public List<Portlet> getDataSiteLevelPortlets(
+			long companyId, boolean excludeDataAlwaysStaged)
+		throws Exception {
+
+		return getExportImportHelper().getDataSiteLevelPortlets(
+			companyId, excludeDataAlwaysStaged);
+	}
+
 	/**
 	 * @deprecated As of 7.0.0, moved to {@link
 	 *             ExportImportDateUtil#getDateRange(PortletRequest, long,
